naming change fix

+++ src/database/mod.rs
@@ -341,9 +341,9 @@ impl Database {
         Ok(result
             .iter()
             .map(|row| {
-                let batch_size = row.get::<i64, _>(0) as u64;
+                let batch_size = row.get::<i64, _>(0);
                 let url = row.get::<String, _>(1);
-                let timeout_s = row.get::<i64, _>(2) as u64;
+                let timeout_s = row.get::<i64, _>(2);
                 prover::Prover {
                     url,
                     batch_size,
